requirements.txt
python-telegram-bot
bs4
requests